feat: share button, help modal, how-it-works landing page, community branding
- Add Share button (copies URL, flips to Copied confirmation)
- Add ? help modal with 3-step guide and share CTA
- Home page: 4-step 'How it works' section below hero
- Login: rebrand from Metacrisis Salon 9 to Community Scheduling Assistant
- Event page: sticky sidebar, desktop inline inspect card
- Calendar: full-width responsive fix (align-items: stretch in column layout)
- HourBox: selected cell inset border for edit-mode clarity

+++ client/src/utils.ts
@@ -107,7 +107,7 @@ const getDay = function (unixObject: number) {
     return "";
   }
   const a = new Date(unixObject * 1000);
-  return Intl.DateTimeFormat("en-US", { weekday: "long" }).format(a);
+  return Intl.DateTimeFormat("en-US", { weekday: "short" }).format(a);
 };
 
 export {
